relay/stream_manager/requirements.txt
awscrt==0.23.5
awsiotsdk==1.22.1
dotenv==0.9.9
python-dotenv==1.0.1
